+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Fix detection of Linux ARM-based Android. See [3b0e113c8047](https://github.com/mssola/user_agent/commit/3b0e113c804708c01de00c27aae07d2acfee40d8).
 - Add detection of Chromium Edge on Windows. See [ea81f1e9d61c](https://github.com/mssola/user_agent/commit/ea81f1e9d61c094df4156690a8f4d5481b0d6c4a).
 - Add detection of OkHttp. See [6b33e248e796](https://github.com/mssola/user_agent/commit/6b33e248e7969cf3e76128a34d33be88d4eb0dc8).
+- Add detection of Coc Coc Browser.
+- Add detection of Headless Chrome.
+- Add detection of iOS WebViews.
 
 ## 0.5.2